+++ branches/2.0/assets/metabox/driver/icon/js/scripts.js
@@ -1,3 +1,3 @@
 'use strict';
 
-import 'presstify-framework/field/select-image/js/scripts';
+import '../../../../field/select-image/js/scripts';