+++ branches/2.0/assets/partial/flash-notice/js/scripts.js
@@ -1,4 +1,4 @@
 'use strict';
 
 // Dépendances
-import 'presstify-framework/partial/notice/js/scripts';
+import '../../../partial/notice/js/scripts';