+++ branches/2.0/src/Form/Resources/views/body.php
@@ -3,11 +3,11 @@
  * Corps du formulaire.
  * ---------------------------------------------------------------------------------------------------------------------
  * @var tiFy\Contracts\Form\FactoryView $this
- * @var tiFy\Contracts\Form\FactoryFields|tiFy\Contracts\Form\FactoryFields $fields
+ * @var tiFy\Contracts\Form\FactoryFields $fields
  */
 ?>
 <?php if ($fields->exists()) : ?>
-    <div class="Form-fields">
+    <div class="FormFields">
         <?php $this->insert('groups', $this->all()); ?>
     </div>
 <?php endif; ?>